+++ CHANGES.rst
@@ -15,6 +15,7 @@ in progress
 - Primary key detection
 - Content type detection using file extension or ``--content-type`` option
 - Add support for CSV input format
+- Add support for XLSX and ODS input formats
 
 
 2023-01-xx 0.0.0